make `darkMode: 'media'`, the default
This also includes moving dark mode tests to its own dedicated file.
