+++ integrations/postcss-cli/tests/integration.test.js
@@ -111,7 +111,6 @@ describe('watcher', () => {
           module.exports = {
             purge: ['./src/index.html'],
             mode: 'jit',
-            darkMode: false, // or 'media' or 'class'
             theme: {
               extend: {
                 screens: {
