+++ integrations/tailwindcss-cli/tests/cli.test.js
@@ -117,7 +117,6 @@ describe('Build command', () => {
       {
         purge: ['./src/index.html'],
         mode: 'jit',
-        darkMode: false, // or 'media' or 'class'
         theme: {
           extend: {
             fontWeight: {
