+++ integrations/vite/tailwind.config.js
@@ -1,7 +1,6 @@
 module.exports = {
   purge: ['./index.html'],
   mode: 'jit',
-  darkMode: false, // or 'media' or 'class'
   theme: {
     extend: {},
   },
