+++ integrations/vite/tests/integration.test.js
@@ -134,7 +134,6 @@ describe('watcher', () => {
         module.exports = {
           purge: ['./index.html'],
           mode: 'jit',
-          darkMode: false, // or 'media' or 'class'
           theme: {
             extend: {
               screens: {
