+++ stubs/simpleConfig.stub.js
@@ -1,6 +1,6 @@
 module.exports = {
   purge: [],
-  darkMode: false, // or 'media' or 'class'
+  darkMode: 'media', // or 'class'
   theme: {
     extend: {},
   },
